validation improvements
