+++ src/PH/Bundle/CoreBundle/Form/Extension/MollieGatewayConfigurationExtensionType.php
@@ -5,41 +5,9 @@ declare(strict_types=1);
 namespace PH\Bundle\CoreBundle\Form\Extension;
 
 use PH\Bundle\PayumBundle\Form\Type\MollieGatewayConfigurationType;
-use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\Extension\Core\Type\IntegerType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Validator\Constraints\NotBlank;
-use Symfony\Component\Validator\Constraints\Range;
 
-final class MollieGatewayConfigurationExtensionType extends AbstractTypeExtension
+final class MollieGatewayConfigurationExtensionType extends AbstractGatewayConfigurationExtension
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('minAmount', IntegerType::class, [
-                'constraints' => [
-                    new NotBlank([
-                        'groups' => 'ph',
-                    ]),
-                    new Range([
-                        'min' => 0,
-                        'groups' => 'ph',
-                    ]),
-                ],
-            ])
-            ->add('maxAmount', IntegerType::class, [
-                'constraints' => [
-                    new NotBlank([
-                        'groups' => 'ph',
-                    ]),
-                ],
-            ])
-        ;
-    }
-
     /**
      * {@inheritdoc}
      */
